requirements.txt
# Core quantum + autodiff
pennylane>=0.36
jax>=0.4.25
jaxlib>=0.4.25

# Numerical computing
numpy>=1.24
scipy>=1.10

# Machine learning + dimensionality reduction
scikit-learn>=1.3
umap-learn>=0.5.5

# Data handling
pandas>=2.0

# Plotting
matplotlib>=3.7
seaborn>=0.12

# Optional but useful (progress + experiments)
tqdm>=4.66